+++ UVtools.Core/Managers/IssueManager.cs
@@ -218,7 +218,7 @@ namespace UVtools.Core.Managers
 
                     using (var image = layer.LayerMat)
                     {
-                        int step = image.Step;
+                        var step = image.GetRealStep();
                         var span = image.GetDataByteSpan();
 
                         if (touchBoundConfig.Enabled)
@@ -396,7 +396,7 @@ namespace UVtools.Core.Managers
                                         if (previousImage is null)
                                         {
                                             previousImage = SlicerFile[layerIndex - 1].LayerMat;
-                                            previousSpan = previousImage.GetDataSpan<byte>();
+                                            previousSpan = previousImage.GetDataByteSpan();
                                         }
 
                                         List<Point> points = new();
@@ -463,10 +463,11 @@ namespace UVtools.Core.Managers
                                                 anchor, overhangConfig.ErodeIterations, BorderType.Default,
                                                 new MCvScalar());
 
-                                            var subtractedSpan = subtractedImage.GetDataSpan<byte>();
+                                            var subtractedSpan = subtractedImage.GetDataByteSpan();
+                                            var subtractedStep = subtractedImage.GetRealStep();
 
                                             for (int y = 0; y < subtractedImage.Height; y++)
-                                                for (int x = 0; x < subtractedImage.Step; x++)
+                                                for (int x = 0; x < subtractedStep; x++)
                                                 {
                                                     int labelX = rect.X + x;
                                                     int labelY = rect.Y + y;
@@ -849,10 +850,9 @@ namespace UVtools.Core.Managers
                                     var overlappingGroupIndexes = new List<int>();
                                     for (var groupIndex = 0; groupIndex < resinTrapGroups.Count; groupIndex++)
                                     {
-                                        if (resinTrapGroups[groupIndex].Last().layerIndex != layerIndex + 1) continue;
+                                        if (resinTrapGroups[groupIndex][^1].layerIndex != layerIndex && resinTrapGroups[groupIndex][^1].layerIndex != layerIndex + 1) continue;
 
-                                        if (EmguContours.ContoursIntersect(resinTrapGroups[groupIndex].Last().contour,
-                                            resinTraps[layerIndex][x]))
+                                        if (EmguContours.ContoursIntersect(resinTrapGroups[groupIndex][^1].contour, resinTraps[layerIndex][x]))
                                         {
                                             overlappingGroupIndexes.Add(groupIndex);
                                         }
@@ -877,8 +877,8 @@ namespace UVtools.Core.Managers
 
                                         for (var index = overlappingGroupIndexes.Count - 1; index >= 0; index--)
                                         {
-                                            resinTrapGroups[index].Clear();
-                                            resinTrapGroups.RemoveAt(index);
+                                            resinTrapGroups[overlappingGroupIndexes[index]].Clear();
+                                            resinTrapGroups.RemoveAt(overlappingGroupIndexes[index]);
                                         }
 
                                         combinedGroup.Add((resinTraps[layerIndex][x], (uint)layerIndex));
@@ -961,9 +961,9 @@ namespace UVtools.Core.Managers
                             var overlappingGroupIndexes = new List<int>();
                             for (var x = 0; x < resinTrapGroups.Count; x++)
                             {
-                                if (resinTrapGroups[x].Last().LayerIndex != layerIndex + 1) continue;
+                                if (resinTrapGroups[x][^1].LayerIndex != layerIndex && resinTrapGroups[x][^1].LayerIndex != layerIndex + 1) continue;
 
-                                using var vec = new VectorOfVectorOfPoint(resinTrapGroups[x].Last().Contours);
+                                using var vec = new VectorOfVectorOfPoint(resinTrapGroups[x][^1].Contours);
                                 if (EmguContours.ContoursIntersect(trap, vec))
                                 {
                                     overlappingGroupIndexes.Add(x);
@@ -989,8 +989,8 @@ namespace UVtools.Core.Managers
 
                                 for (var index = overlappingGroupIndexes.Count - 1; index >= 0; index--)
                                 {
-                                    resinTrapGroups[index].Clear();
-                                    resinTrapGroups.RemoveAt(index);
+                                    resinTrapGroups[overlappingGroupIndexes[index]].Clear();
+                                    resinTrapGroups.RemoveAt(overlappingGroupIndexes[index]);
                                 }
 
                                 combinedGroup.Add(trapIssue);
@@ -1031,7 +1031,7 @@ namespace UVtools.Core.Managers
                                 var overlappingGroupIndexes = new List<int>();
                                 for (var x = 0; x < suctionGroups.Count; x++)
                                 {
-                                    if (suctionGroups[x][^1].LayerIndex != layerIndex + 1) continue;
+                                    if (suctionGroups[x][^1].LayerIndex != layerIndex && suctionGroups[x][^1].LayerIndex != layerIndex + 1) continue;
                                     using var vec = new VectorOfVectorOfPoint(suctionGroups[x][^1].Contours);
                                     if (EmguContours.ContoursIntersect(trap, vec))
                                     {